Moved metamask handling to Step 3
